requirements.txt
numpy==2.0.1
imutils~=0.5.4
pillow==10.4.0
pandas==2.2.2
requests==2.32.3
fastapi[standard]==0.112.1
opencv-python==4.10.0.84
dlib==19.24.5
face-recognition==1.3.0
scikit-learn==1.5.1
torch==2.4.0
fastapi==0.112.1